feat: add test webhook button for verifying new webhook setups
Each configured webhook now has a paperplane button that sends a
sample "Status Degraded" event so users can verify delivery without
waiting for a real incident. Extracted WebhooksSettingsTab to its own
file to stay within SwiftLint's 600-line file length limit.
